Add LoadGameUseCase

+++ app/src/main/java/com/piticlistudio/playednext/data/game/model/GameModel.kt
@@ -4,4 +4,4 @@ package com.piticlistudio.playednext.data.game.model
  * Representation for a [GameModel] fetched for the data layer
  */
 class GameModel(val id: Int, val name: String, val summary: String?, val storyline: String?,
-                val collection: Int?, val franchise: Int?, val rating: Int?)
+                val collection: Int?, val franchise: Int?, val rating: Float?)
